fix: Resolve all compilation errors (Long/Integer and Optional type mismatches)

+++ src/main/java/com/abhi/authProject/controller/AdminUserController.java
@@ -79,7 +79,7 @@ public class AdminUserController {
     }
 
     @DeleteMapping("/users/{id}")
-    public ResponseEntity<?> deleteUser(@PathVariable Long id) {
+    public ResponseEntity<?> deleteUser(@PathVariable Integer id) {
         if (userRepo.existsById(id)) {
             userRepo.deleteById(id);
             return ResponseEntity.ok().build();
